Disable swap() in safe mode, revert errors, view functions
When hotPathOpen_ && inSafeMode_ it was possible to execute a swap on
CrocSwapDex but not on any proxy paths. For safety this has been
disabled and will return a "LK" revert message to indicate that the DEX
is locked.

Several common revert situations now return an error string to aid in
debugging.

The authority_ and inSafeMode_ variables are exposed through public view
authority() and safeMode() functions, respectively