Add events to common operations (swap, mint, etc.)
Added InitPool, Swap MintRanged, MintAmbient, BurnRanged, BurnAmbient, and
Harvest events to CrocEvents and inserted them in the Hot, Cold, Warm,
and Long paths where they take place. In the case of LongPath it is
possible for many events to happen on the same block, one event should
be created for each swap, mint, burn, etc.