Remove Timelock.sol and timelock specific tests
This patch removes the now unused Timelock.sol contract but does not
remove the mocktimelock contract that features heavily on the governance
tests.

While functionally it shouldn't make a difference the proper way to do
this would be to refactor the CrocPolicy and PoolPolicy tests to not use
the mocktimelock contract and work from there.