Add liq param to knockout events
This param is needed by the backend to facilitate position tracking.
While it is possible to reproduce the liq value by calculating it from
the base flow, quote flow, and current price it is far easier to receive
the value directly from the source of truth and make use of it from
there.